refactor: #601 상품 조회/명령 책임 분리로 유지보수 충돌을 줄인다
ProductsService를 얇은 파사드로 축소하고 읽기 책임은 ProductQueryService, 쓰기 책임은 ProductCommandService로 분리했습니다.
create/update의 이미지·상세이미지 처리 흐름은 upsertMediaEntities 공통 helper로 통합해 정렬/썸네일 규칙을 한 경로에서 유지하도록 정리했습니다.
캐시 무효화도 invalidateProductCache 단일 메서드로 묶어 detail/list 패턴 삭제 규칙을 일관화했습니다.

Constraint: 기존 ProductsService 공개 메서드 시그니처와 컨트롤러 계약을 유지해야 함
Rejected: 컨트롤러를 Query/Command 서비스로 직접 분기 주입 | 호출부 변경 범위가 커지고 회귀 리스크 증가
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: 상품 도메인 규칙 추가 시 조회 로직은 ProductQueryService, 쓰기/부수효과는 ProductCommandService로만 확장할 것
Tested: cd backend && npm run test -- products.service.spec.ts products.search.spec.ts
Tested: cd backend && npm run build && npm run test
Tested: npm run build && npm run test:run
Not-tested: 실제 관리자 화면에서 상품 생성/수정 UI를 통한 수동 시나리오
