+++ backend/src/modules/products/__tests__/products.search.spec.ts
@@ -3,6 +3,8 @@ import { getRepositoryToken, getDataSourceToken } from '@nestjs/typeorm';
 import { BadRequestException } from '@nestjs/common';
 import { DataSource, SelectQueryBuilder, QueryFailedError } from 'typeorm';
 import { ProductsService } from '../products.service';
+import { ProductQueryService } from '../product-query.service';
+import { ProductCommandService } from '../product-command.service';
 import { Product, ProductStatus } from '../entities/product.entity';
 import { Category } from '../entities/category.entity';
 import { Review } from '../../reviews/entities/review.entity';
@@ -69,6 +71,8 @@ describe('ProductsService — Search', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         ProductsService,
+        ProductQueryService,
+        ProductCommandService,
         {
           provide: getRepositoryToken(Product),
           useValue: mockRepository,
