+++ backend/src/modules/products/products.module.ts
@@ -9,6 +9,8 @@ import { Review } from '../reviews/entities/review.entity';
 import { AttributeType } from './entities/attribute-type.entity';
 import { ProductAttribute } from './entities/product-attribute.entity';
 import { ProductsService } from './products.service';
+import { ProductQueryService } from './product-query.service';
+import { ProductCommandService } from './product-command.service';
 import { CategoriesService } from './categories.service';
 import { AttributesService } from './attributes.service';
 import { RecentlyViewedService } from './recently-viewed.service';
@@ -35,8 +37,22 @@ import { RecentlyViewedProduct } from './entities/recently-viewed-product.entity
     CacheModule,
     RestockAlertsModule,
   ],
-  providers: [ProductsService, CategoriesService, AttributesService, RecentlyViewedService],
+  providers: [
+    ProductsService,
+    ProductQueryService,
+    ProductCommandService,
+    CategoriesService,
+    AttributesService,
+    RecentlyViewedService,
+  ],
   controllers: [ProductsController, CategoriesController, AttributesController],
-  exports: [ProductsService, CategoriesService, AttributesService, RecentlyViewedService],
+  exports: [
+    ProductsService,
+    ProductQueryService,
+    ProductCommandService,
+    CategoriesService,
+    AttributesService,
+    RecentlyViewedService,
+  ],
 })
 export class ProductsModule {}
